LM3S6432-S2E: Rename files for better conformance to naming conventions
