+++ nuttx/configs/lm3s6432-s2e/src/lm3s6432s2e_internal.h
@@ -47,7 +47,7 @@
 #include "tiva_gpio.h"
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 
 /* How many SSI modules does this chip support? The LM3S6432 supports 1 SSI
