+++ nuttx/configs/lm3s6432-s2e/src/lm_ethernet.c
@@ -1,6 +1,5 @@
 /************************************************************************************
- * configs/lm3s6432-s3e/src/up_ethernet.c
- * arch/arm/src/board/up_ethernet.c
+ * configs/lm3s6432-s3e/src/lm_ethernet.c
  *
  *   Copyright (C) 2010 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
