+++ nuttx/configs/lm3s6432-s2e/src/lm_nsh.c
@@ -1,6 +1,5 @@
 /****************************************************************************
- * config/lm3s6432-s2e/src/up_nsh.c
- * arch/arm/src/board/up_nsh.c
+ * config/lm3s6432-s2e/src/lm_nsh.c
  *
  *   Copyright (C) 2010 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -48,7 +47,7 @@
 #include <nuttx/mmcsd.h>
 
 /****************************************************************************
- * Pre-Processor Definitions
+ * Pre-processor Definitions
  ****************************************************************************/
 
 /****************************************************************************
